third_party/blink/web_tests/platform/generic/html/tabular_data/tbody_rows-expected.txt

Test behavior of the HTMLTableSectionElement rows attribute in cases where there is unusual nesting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkRowNesting("col") is 0
PASS checkRowNesting("colgroup") is 0
PASS checkRowNesting("div") is 0
PASS checkRowNesting("form") is 0
PASS checkRowNesting("script") is 0
PASS checkRowNesting("table") is 0
PASS checkRowNesting("td") is 0
PASS checkRowNesting("th") is 0


PASS checkRowNesting("tbody") is 0
PASS checkRowNesting("tfoot") is 0
PASS checkRowNesting("thead") is 0


PASS checkRowNesting("tr") is 1


PASS successfullyParsed is true

TEST COMPLETE
